requirements.txt
asgiref==3.3.4
cloudinary==1.25.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-filter==23.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.2.0
PyJWT==2.8.0
sqlparse==0.4.1
urllib3==1.26.16
